Arithmetic fixes, improved a few extracodes. (#10)
NEGATE is now a routine.
Arith: corrected operations with -1.
Updated E71 and E50 130.
Corrected elementary functions for negative values of the argument.
Catering to compilers with no __int128 support.
Using proper ways to convert BESM6 <-> IEEE and to check for denormalized values.
WHATIS is now satisfied with the results of E50 0130.
Added "-e" for native elem. functions.
Added implementations of the elementary functions.
Implemented capturing of plotter output for "ГОРОСКОП" (Э50 1211).
Corrected E75 in presence of 'SPEC'.